requirements.txt
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
SQLAlchemy==2.0.13
psycopg2-binary==2.9.6
black==23.3.0
pylint==2.17.4
